drafting "Minimax Search and ...", incl. pilot images
I wanted to use a cool hackerish version-controllable presentation
framework, but with the conference but three weeks away (and with
travel and the Rust All-Hands in the meantime), it might be more
time-efficient to just use boring ol' LibreOffice Impress. But I
mostly want the slides to be images anyway (no one wants to read
bullet points!); I can check those in, and worry about slapping them
into presentation software later.