"Minimax Search and ...": first contiguous script draft
This, at last, makes me feel less nervous. Having written up a
transcript of what I want to say, all that's left is to practice
(practice, practice) saying it, and put together the slide deck. Doing
that in the week and a half before the conference doesn't even seem
that unreasonable.